Refaktorering og logging
