+++ gradle.properties
@@ -14,5 +14,5 @@ exposedVersion=0.55.0
 flywayCoreVersion=10.20.0
 hikariVersion=6.0.0
 postgresqlVersion=42.7.4
-kotlinterVersion=4.4.0
+kotlinterVersion=4.4.1
 h2_version=2.3.232
