+++ src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/Inntektsmelding.kt
@@ -1,0 +1,14 @@
+package no.nav.helsearbeidsgiver.inntektsmelding
+
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonObject
+
+@Serializable
+data class Inntektsmelding(
+    val dokument: JsonObject,
+    val orgnr: String,
+    val fnr: String,
+    val foresporselid: String?,
+    val innsendt: String,
+    val mottattEvent: String,
+)
