+++ src/test/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/HentInntektsmeldingerApiTest.kt
@@ -11,7 +11,7 @@ class HentInntektsmeldingerApiTest :
             testApi {
                 val response = get("/inntektsmeldinger")
                 response.status.value shouldBe 200
-                val inntektsmeldinger = response.body<List<ExposedInntektsmelding>>()
+                val inntektsmeldinger = response.body<List<Inntektsmelding>>()
                 inntektsmeldinger.size shouldBe 1
                 inntektsmeldinger[0].orgnr shouldBe "810007842"
                 inntektsmeldinger[0].dokument["vedtaksperiodeId"] shouldBe "13129b6c-e9f5-4b1c-a855-abca47ac3d7f".toJson()
